Fix a few warnings flagged by a MinGW toolchain (#490)
* Fix a few warnings flagged by a MinGW toolchain

* Fix MenuWidget::DrawString

* clang-format

* cmake: create MAP file when using MSVC toolchains

+++ LEGORacers/src/race/slatebridge0x68.cpp
@@ -149,8 +149,8 @@ void SlateBridge0x68::FUN_00427440()
 			0xffffff00 -
 			static_cast<LegoS32>(static_cast<LegoFloat>(m_unk0x54) * 0.002f * g_twoPi * g_negativeRadiansToTableIndex);
 		index &= c_cosineTableMask;
-		selectedAlpha =
-			c_selectedPulseAlphaCenter - static_cast<LegoS32>(g_cosineTable[index] * c_selectedPulseAlphaAmplitude);
+		selectedAlpha = c_selectedPulseAlphaCenter -
+						static_cast<LegoS32>(g_cosineTable[index] * (int) c_selectedPulseAlphaAmplitude);
 	}
 
 	m_font->SetColor(c_promptColor);
